Rename component template API route handler files to follow pattern used elsewhere.
