+++ x-pack/plugins/index_management/server/routes/api/component_templates/index.ts
@@ -7,11 +7,11 @@
 
 import { RouteDependencies } from '../../../types';
 
-import { registerGetAllRoute } from './get';
-import { registerCreateRoute } from './create';
-import { registerUpdateRoute } from './update';
-import { registerDeleteRoute } from './delete';
-import { registerPrivilegesRoute } from './privileges';
+import { registerGetAllRoute } from './register_get_route';
+import { registerCreateRoute } from './register_create_route';
+import { registerUpdateRoute } from './register_update_route';
+import { registerDeleteRoute } from './register_delete_route';
+import { registerPrivilegesRoute } from './register_privileges_route';
 
 export function registerComponentTemplateRoutes(dependencies: RouteDependencies) {
   registerGetAllRoute(dependencies);
