+++ x-pack/plugins/index_management/server/routes/api/component_templates/register_create_route.ts
