+++ x-pack/plugins/index_management/server/routes/api/component_templates/register_delete_route.ts
