+++ x-pack/plugins/index_management/server/routes/api/component_templates/register_get_route.ts
