+++ x-pack/plugins/index_management/server/routes/api/component_templates/register_privileges_route.ts
