+++ x-pack/plugins/index_management/server/routes/api/component_templates/register_update_route.ts
